Actually fix replay buffer caching and fix state transforms
Previously, e.g. MinMaxScaling would re-derive the min and max values on
the train and val sets separately. Instead, now we train the transform
on the train set, store it in the corresponding checkpoint directory,
and also use the same trained transforms for the val set.

This probably did not create any issues for most variables, as the key
ones are bounded (INR and age).