Update preprocessing script
Ensure that non-splittable events that we include don't mess up the
dynamics where they occur. Identify an issue in the computation of
`not_done`, specifically identifying the last transition in a trajectory
does not seem to work. However, it raises questions whether this is the
right approach when a trajectory likely continues unobserved.