Naive implementation of new interruptions data
Split trajectories in all trials on the new `INTERRUPTION_FLAG` prior to
doing trial-specific preprocessing.

This significantly shortens average trajectory length in most studies.

What remains is to identify whether some trial-specific splitting is
still needed, or if it would be better to exclude trajectories we were
previously splitting on the basis that we do not know if our heuristic
approach to identifying interruptions is still valid.