Allow using pre-trained behavior cloner for BCQ
Initialize the "generative network" (in terms of our paper, the model
that determines the "feasible set" of options to take) from a
pre-trained behavior cloner model, which should improve the rate of
convergence during BCQ training. By default, this can also be frozen,
which should help if/when we do event trajectory oversampling
subanalyses.

Fixes #97