Add code for removing unintuitive decisions
Add script that removes transitions where the observed clinical decision
is not consistent with the current INR (i.e. increasing when INR high,
decreasing when INR low, with some additional complexity around dose
resumption).

Fix the auditing helper so that when we prune short trajectories
afterward we don't overwrite the auditing of the main call.

Fixes #103