Work on pulling forward data across trajectories
Adds the ability to pull INR history forward from across trajectories,
with the goal of adding more information to the state space. Also adds
the ability to pull forward past doses. The concern there will be
violations of the Markov assumptions, but I think there are obvious
ameliorations if it turns out to provide any performance improvement.

Working on #106