Update evaluation pipeline with two tuning metrics
Add a weighted least squares regression that predicts TTR from
trajectory-level algorithm consistency. Add a weighted logistic
regression that predicts event occurrence from trajectory-level
algorithm-consistency.

Remove out-of-date power analysis.

This commit now represents the location of the current best experiment,
as all it changes is the evaluation pipeline, and not in a way that
affects the test procedure, which does not rely on either of these
tuning metrics.

Fixes #118, fixes #122